requirements.txt
Django==4.0.1
djangorestframework==3.13.1
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
pep8-naming==0.12.1